+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-alpha.28
+
+### Patch Changes
+
+- feat: compNode init after constructor
+
 ## 1.0.0-alpha.27
 
 ### Patch Changes